+++ src/frontend/packages/lib_classroom/src/components/ClassroomWidgetProvider/widgets/Recordings/index.spec.tsx
@@ -1,5 +1,10 @@
 import { screen } from '@testing-library/react';
-import { InfoWidgetModalProvider } from 'lib-components';
+import fetchMock from 'fetch-mock';
+import {
+  InfoWidgetModalProvider,
+  useCurrentResourceContext,
+  useJwt,
+} from 'lib-components';
 import { render } from 'lib-tests';
 import { DateTime } from 'luxon';
 import React from 'react';
@@ -12,8 +17,35 @@ import { wrapInClassroom } from '@lib-classroom/utils/wrapInClassroom';
 
 import { Recordings } from '.';
 
+jest.mock('lib-components', () => ({
+  ...jest.requireActual('lib-components'),
+  report: jest.fn(),
+  useCurrentResourceContext: jest.fn(),
+}));
+
+const mockedUseCurrentResourceContext =
+  useCurrentResourceContext as jest.MockedFunction<
+    typeof useCurrentResourceContext
+  >;
+
 describe('<Recordings />', () => {
+  beforeEach(() => {
+    useJwt.setState({
+      jwt: 'token',
+    });
+  });
+
+  afterEach(() => {
+    fetchMock.restore();
+    jest.resetAllMocks();
+  });
+
   it('displays a list of available recordings', () => {
+    mockedUseCurrentResourceContext.mockReturnValue([
+      {
+        isFromWebsite: false,
+      },
+    ] as any);
     let classroom = classroomMockFactory({ id: '1', started: false });
     const classroomRecordings = [
       classroomRecordingMockFactory({
@@ -62,6 +94,16 @@ describe('<Recordings />', () => {
     ).toBeInTheDocument();
     expect(
       screen.getByText('Tuesday, February 15, 2022 - 11:00 AM'),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {
+        name: 'Convert Tuesday, March 1, 2022 - 11:00 AM to VOD',
+      }),
+    ).toBeInTheDocument();
+    expect(
+      screen.getByRole('button', {
+        name: 'Convert Tuesday, February 15, 2022 - 11:00 AM to VOD',
+      }),
     ).toBeInTheDocument();
   });
 });